packages/shared-reporting/pyproject.toml
[project]
name = "growthnav-reporting"
version = "0.1.0"
description = "Shared reporting framework with PDF, Google Sheets, and Google Slides output"
requires-python = ">=3.11"
dependencies = [
    "jinja2>=3.1.0",
    "weasyprint>=60.0",
    "gspread>=6.0.0",
    "google-auth>=2.0.0",
    "google-api-python-client>=2.0.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["growthnav"]
